Split large arrays into segments
When opening a large array, the UI gets paused for a long time while the
new DOM elements are created. If this array has thousands of elements,
that can take a long time. It probably isn't useful to render thousands
of things (a person will probably only look at a few of them), so
splitting them into chunks of 100 is probably reasonable. The chunks
could probably be even smaller, but this is fine for now.